# backend/openapi/k12-api-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "K12 平台接口文档",
    "description": "K12 多智能体教学平台可调用接口。包含 IAM 账号安全、课程、智能体、结构化作业题目、答题与批改，统一通过 Gateway 和 JWT 鉴权。",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "网关服务"
    },
    {
      "url": "http://localhost:8081",
      "description": "用户与权限服务"
    },
    {
      "url": "http://localhost:8082",
      "description": "课程资源服务"
    },
    {
      "url": "http://localhost:8083",
      "description": "AI 智能体服务"
    },
    {
      "url": "http://localhost:8084",
      "description": "作业与批改服务"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/learning/courses/{courseId}/chapters": {
      "get": {
        "tags": [
          "02-课程资源服务"
        ],
        "summary": "查询课程章节目录",
        "description": "需要 course:read 或管理员，且为课程创建人、管理员或已报名学生。仅返回目录，不返回正文。",
        "operationId": "listCourseChapters",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterListApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "403": {
            "description": "功能权限或数据权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "404": {
            "description": "课程、章节或报名记录不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "409": {
            "description": "资源冲突或章节数量达到上限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "02-课程资源服务"
        ],
        "summary": "新增课程章节",
        "description": "需要 course:update 且为课程创建人，或管理员。每门课程最多 500 个有效章节。",
        "operationId": "createCourseChapter",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "201": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "403": {
            "description": "功能权限或数据权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "404": {
            "description": "课程、章节或报名记录不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "409": {
            "description": "资源冲突或章节数量达到上限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/learning/courses/{courseId}/chapters/{chapterId}": {
      "get": {
        "tags": [
          "02-课程资源服务"
        ],
        "summary": "查询章节正文",
        "description": "权限同章节目录，章节必须属于指定课程。",
        "operationId": "getCourseChapter",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "chapterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "403": {
            "description": "功能权限或数据权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "404": {
            "description": "课程、章节或报名记录不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "409": {
            "description": "资源冲突或章节数量达到上限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "02-课程资源服务"
        ],
        "summary": "修改章节内容与顺序",
        "description": "需要 course:update 且为课程创建人，或管理员。",
        "operationId": "updateCourseChapter",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "chapterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "403": {
            "description": "功能权限或数据权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "404": {
            "description": "课程、章节或报名记录不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "409": {
            "description": "资源冲突或章节数量达到上限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterRequest"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "02-课程资源服务"
        ],
        "summary": "删除课程章节",
        "description": "需要 course:update 且为课程创建人，或管理员。逻辑删除后章节不再计入学习进度。",
        "operationId": "deleteCourseChapter",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "chapterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "400": {
            "description": "参数不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "403": {
            "description": "功能权限或数据权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "404": {
            "description": "课程、章节或报名记录不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          },
          "409": {
            "description": "资源冲突或章节数量达到上限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBase"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/learning/courses/{id}": {
      "get": {
        "tags": [
          "02-课程资源服务"
        ],
        "summary": "课程详情",
        "operationId": "getCourse",
        "description": "需要 ROLE_ADMIN 或 course:read 权限。",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "课程详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseApiResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "02-课程资源服务"
        ],
        "summary": "更新课程",
        "operationId": "updateCourse",
        "description": "需要 course:update 且为课程创建人，或 ROLE_ADMIN。无归属的历史课程仅允许管理员维护。",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/CourseRequest"
        },
        "responses": {
          "200": {
            "description": "更新后的课程",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "02-课程资源服务"
        ],
        "summary": "删除课程",
        "operationId": "deleteCourse",
        "description": "需要 course:delete 且为课程创建人，或 ROLE_ADMIN。逻辑删除课程，保留历史报名与进度。",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkEmpty"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "example": 1
        }
      }
    },
    "requestBodies": {
...
      "CourseRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CourseRequest"
            },
            "example": {
              "title": "一次函数基础",
              "subject": "数学",
              "gradeLevel": "八年级",
              "description": "学习一次函数的定义和图像"
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "未登录、JWT 无效或 JWT 已过期",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            },
            "example": {
              "code": 401,
              "message": "未登录、登录已过期或令牌无效",
              "data": null,
              "timestamp": "2026-08-14T02:30:00Z"
            }
          }
        }
      },
      "Forbidden": {
        "description": "已登录，但没有访问该资源的权限",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            },
            "example": {
              "code": 403,
              "message": "没有访问该资源的权限",
              "data": null,
              "timestamp": "2026-08-14T02:30:00Z"
            }
          }
        }
      },
      "BadRequest": {
        "description": "请求参数或业务输入不合法",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      },
      "OkEmpty": {
        "description": "操作成功",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EmptyApiResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiResponseBase": {
        "type": "object",
        "required": [
          "code",
          "message",
          "timestamp"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EmptyApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "nullable": true
              }
            }
          }
        ]
      },
...
      "CourseRequest": {
        "type": "object",
        "required": [
          "title",
          "subject",
          "gradeLevel"
        ],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 128,
            "example": "一次函数基础"
          },
          "subject": {
            "type": "string",
            "maxLength": 64,
            "example": "数学"
          },
          "gradeLevel": {
            "type": "string",
            "maxLength": 32,
            "example": "八年级"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "nullable": true,
            "example": "学习一次函数的定义和图像"
          }
        }
      },
...
      "ChapterRequest": {
        "type": "object",
        "required": [
          "title",
          "content",
          "sortOrder"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "example": "第一章：一次函数"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000,
            "description": "纯文本正文；前端按文本展示，不作为 HTML 执行",
            "example": "本章学习一次函数的定义。"
          },
          "sortOrder": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10000,
            "example": 1
          }
        }
      },
      "ChapterSummary": {
        "type": "object",
        "required": [
          "id",
          "courseId",
          "title",
          "sortOrder",
          "updatedTime"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "courseId": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "title": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Chapter": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ChapterSummary"
          },
          {
            "type": "object",
            "required": [
              "content"
            ],
            "properties": {
              "content": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "ChapterApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Chapter"
              }
            }
          }
        ]
      },
...
      "ChapterListApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ChapterSummary"
                }
              }
            }
          }
        ]
      },
      "Course": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CourseRequest"
          },
          {
            "type": "object",
            "properties": {
              "teacherId": {
                "type": "integer",
                "format": "int64",
                "nullable": true,
                "description": "课程创建人的用户 ID；历史课程可能为空"
              },
              "id": {
                "type": "integer",
                "format": "int64",
                "example": 1
              },
              "updatedTime": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "CourseApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponseBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Course"
              }
            }
          }
        ]
      },
...
    }
  }
}